Split stage1 into multiple files
