+++ src/stageinfo.c
@@ -10,7 +10,7 @@
 
 #include "stageinfo.h"
 
-#include "stages/stage1.h"
+#include "stages/stage1/stage1.h"
 #include "stages/stage2.h"
 #include "stages/stage3.h"
 #include "stages/stage4.h"
