+++ src/stages/stage1/background_anim.h
@@ -1,0 +1,21 @@
+/*
+ * This software is licensed under the terms of the MIT License.
+ * See COPYING for further information.
+ * ---
+ * Copyright (c) 2011-2019, Lukas Weber <laochailan@web.de>.
+ * Copyright (c) 2012-2019, Andrei Alexeyev <akari@taisei-project.org>.
+*/
+
+#ifndef IGUARD_stages_stage1_background_anim_h
+#define IGUARD_stages_stage1_background_anim_h
+
+#include "taisei.h"
+
+void stage1_bg_raise_camera(void);
+void stage1_bg_enable_snow(void);
+void stage1_bg_disable_snow(void);
+
+void stage1_bg_init_fullstage(void);
+void stage1_bg_init_spellpractice(void);
+
+#endif // IGUARD_stages_stage1_background_anim_h
