+++ src/stages/stage1/cirno.h
@@ -1,0 +1,20 @@
+/*
+ * This software is licensed under the terms of the MIT License.
+ * See COPYING for further information.
+ * ---
+ * Copyright (c) 2011-2019, Lukas Weber <laochailan@web.de>.
+ * Copyright (c) 2012-2019, Andrei Alexeyev <akari@taisei-project.org>.
+*/
+
+#ifndef IGUARD_stages_stage1_cirno_h
+#define IGUARD_stages_stage1_cirno_h
+
+#include "taisei.h"
+
+#include "boss.h"
+
+Boss *stage1_spawn_cirno(cmplx pos);
+void stage1_draw_cirno_spellbg(Boss *boss, int time);
+void stage1_cirno_wander(Boss *boss, real dist, real lower_bound);
+
+#endif // IGUARD_stages_stage1_cirno_h
