+++ src/stages/stage1/meson.build
@@ -1,0 +1,18 @@
+
+stage1_src = files(
+    'background_anim.c',
+    'cirno.c',
+    'draw.c',
+    'misc.c',
+    'nonspells/boss_nonspell_1.c',
+    'nonspells/boss_nonspell_2.c',
+    'nonspells/midboss_nonspell_1.c',
+    'spells/benchmark.c',  # TODO exclude this conditionally, and maybe move out of stage1
+    'spells/crystal_blizzard.c',
+    'spells/crystal_rain.c',
+    'spells/icicle_cascade.c',
+    'spells/perfect_freeze.c',
+    'spells/snow_halation.c',
+    'stage1.c',
+    'timeline.c',
+)
