+++ src/stages/stage1/misc.h
@@ -1,0 +1,18 @@
+/*
+ * This software is licensed under the terms of the MIT License.
+ * See COPYING for further information.
+ * ---
+ * Copyright (c) 2011-2019, Lukas Weber <laochailan@web.de>.
+ * Copyright (c) 2012-2019, Andrei Alexeyev <akari@taisei-project.org>.
+*/
+
+#ifndef IGUARD_stages_stage1_misc_h
+#define IGUARD_stages_stage1_misc_h
+
+#include "taisei.h"
+
+#include "stage.h"
+
+Projectile *stage1_spawn_stain(cmplx pos, float angle, int to);
+
+#endif // IGUARD_stages_stage1_misc_h
